% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{applyBFS}
\alias{applyBFS}
\title{Baseflow separation}
\usage{
applyBFS(x, d = 5, w = 0.9)
}
\arguments{
\item{x}{numeric vector, values.}

\item{d}{integer, bloc size}

\item{w}{numeric, smoothing parameter}
}
\description{
Apply the baseflow separation (BFS) algorithm described in
Tallaksen and Van Lanenś book (2004):
Hydrological Drought: Processes and Estimation Methods for Streamflow and Groundwater. Elsevier.
}
\examples{
x=rnorm(1001)
bf=applyBFS(x)
plot(x,type='l')
lines(bf,col='red')
}
