% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{applyMovingStat}
\alias{applyMovingStat}
\title{Moving-statistics filter}
\usage{
applyMovingStat(x, window, stat = mean, ...)
}
\arguments{
\item{x}{numeric vector, values.}

\item{window}{numeric, size of the moving window}

\item{stat}{function, function to apply on the moving window}

\item{...}{other arguments passed to stat}
}
\description{
Apply a statistics (typically the mean) on a moving window.
}
\examples{
x=rnorm(100)
y1=applyMovingStat(x,9)
y2=applyMovingStat(x,9,stat=min)
y3=applyMovingStat(x,9,stat=quantile,probs=0.75)
plot(x);lines(y1,col='red');lines(y2,col='blue');lines(y3,col='green')
}
