% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{applySPA}
\alias{applySPA}
\title{Sequent Peak Algorithm (SPA)}
\usage{
applySPA(x, threshold)
}
\arguments{
\item{x}{numeric vector, values.}

\item{threshold}{numeric vector, threshold}
}
\description{
Apply the Sequent Peak Algorithm.
}
\examples{
x=rnorm(100)
spa=applySPA(x,-1)
plot(x,type='l')
lines(rep(-1,100),col='gray')
lines(spa,col='red')
}
