% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distInfo}
\alias{distInfo}
\title{Information on available distributions}
\format{
A named list where each element is itself a list containing:
\describe{
  \item{parName}{parameters short names}
  \item{parLongName}{parameters long names}
  \item{parSymbol}{parameters typical symbols}
  \item{constraints}{constraints on parameters}
  \item{url}{link to more information}
  \item{note}{notes}
  \item{warning}{warnings: read carefully since this highlights in particular differences with "standard" parameterizations found in e.g. Wikipedia or R.}
}
}
\usage{
distInfo
}
\description{
A named list containing information (parameters, contraints, notes, warnings, etc.)
for all available univariate distributions.
}
\keyword{datasets}
