% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_lib.R
\name{Generate}
\alias{Generate}
\title{Random numbers generator}
\usage{
Generate(dist, par, n = 1)
}
\arguments{
\item{dist}{character, distribution name}

\item{par}{numeric vector, parameter vector}

\item{n}{integer, number of values to generate}
}
\value{
The generated values as a numeric vector.
}
\description{
Generate random realizations from a distribution
}
\examples{
Generate('Normal',c(0,1),10)
Generate('GEV',c(100,25,-0.2),10)
Generate('GEV',c(100,25,0.2),10)
Generate('Poisson',0.75,10)
}
