% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_lib.R
\name{GetCdf}
\alias{GetCdf}
\title{Cumulative Distribution Function (cdf)}
\usage{
GetCdf(y, dist, par)
}
\arguments{
\item{y}{numeric, value at which the cdf is evaluated}

\item{dist}{character, distribution name}

\item{par}{numeric vector, parameter vector}
}
\value{
The cdf as a numeric.
}
\description{
Evaluates the cdf of a distribution
}
\examples{
GetCdf(0,'Normal',c(0,1))
GetCdf(200,'GEV',c(100,25,-0.2))
GetCdf(200,'GEV',c(100,25,0.2))
GetCdf(3,'Poisson',0.75)
}
