% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation_HBay.R
\name{GetEstimate_HBay}
\alias{GetEstimate_HBay}
\title{Bayesian estimation of a distribution using historical data}
\usage{
GetEstimate_HBay(
  y,
  dist,
  prior,
  SystErrorIndex = rep(0, NROW(y)),
  SystErrorPrior = list(),
  par0 = GetEstimate_ROUGH(0.5 * (y[, 1] + y[, 2]), dist)$par,
  SystError0 = rep(1, length(SystErrorPrior)),
  mult = 0.1,
  eps = 0.1,
  batch.length = 100,
  batch.n = 100,
  moverate.min = 0.1,
  moverate.max = 0.5,
  mult.down = 0.9,
  mult.up = 1.1
)
}
\arguments{
\item{y}{numeric 2-column matrix, data. 
The first column gives the lower bound, the second column gives the upper bound.
Where y[i,1]==y[i,2], the value is assumed perfectly known (up to systematic errors, see below).
Where y[i,1]<y[i,2], the value is assumed to be in the interval [y[i,1];y[i,2]]
-Inf and +Inf are allowed for data being only right- or left-censored 
(i.e. values known to be smaller than or larger than some threshold).}

\item{dist}{character, distribution name. Only distributions 'GEV', 'Gumbel' and 'LogNormal' are supported.}

\item{prior}{list of lists, prior distributions. For each parameter to be estimated, the prior
is a list of the form pr=list(dist=..., par=...). See example below.}

\item{SystErrorIndex}{integer vector, length NROW(y). Index of systematic errors.
Rows where SystErrorIndex==k are all affected by the same multiplicative error gamma_k, 
typically induced by the kth rating curve. SystErrorIndex==0 means no systematic error.
Should only contain integer values between 0 and N_\{systematic errors\}.}

\item{SystErrorPrior}{list of lists,  prior distribution for each systematic error.
For instance for a systematic error in the range +/- 20\%, you may use a Uniform
between 0.8 and 1.2, or a triangular distribution with the same bounds and peak at 1.}

\item{par0}{numeric vector, initial parameter guess.}

\item{SystError0}{numeric vector, initial guess for systematic errors. Typically a vector of 1.}

\item{mult}{numeric, initial jump standard deviations are set to mult * abs(par0)}

\item{eps}{numeric, where par0 is zero, initial jump standard deviations are set to eps (to avoid jumps of size zero)}

\item{batch.length}{integer, MCMC parameter: length of each non-adaptive batch}

\item{batch.n}{integer, MCMC parameter: number of batches (= adaptation period). Total number of simulations is nsim=batch.n*batch.length}

\item{moverate.min}{numeric in (0;1), MCMC parameter: lower bound for the desired move rate interval}

\item{moverate.max}{numeric in (0;1), MCMC parameter: upper bound for the desired move rate interval}

\item{mult.down}{numeric in (0;1), MCMC parameter: multiplication factor used to decrease jump size when move rate is too low.}

\item{mult.up}{numeric (>1, avoid 1/mult.down), MCMC parameter: multiplication factor used to increase jump size when move rate is too high.}
}
\value{
A list with the following components:
    \item{x}{numeric matrix nsim * (length(par0)+length(SystError0)), MCMC simulations}
    \item{fx}{numeric vector, corresponding values f(x)}
}
\description{
XXX 
Returns MCMC samples from the posterior distribution.
}
