% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_lib.R
\name{GetParFeas}
\alias{GetParFeas}
\title{Parameter feasibility}
\usage{
GetParFeas(dist, par)
}
\arguments{
\item{dist}{character, distribution name}

\item{par}{numeric vector, parameter vector}
}
\value{
A logical.
}
\description{
Evaluates whether a parameter vector is feasible 
(for instance, are scale parameters >0 ?)
}
\examples{
# Feasible
GetParFeas('Normal',c(0,1))
# Not feasible because second parameter (standard deviation) is negative
GetParFeas('Normal',c(0,-1))
}
