% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_lib.R
\name{GetParName}
\alias{GetParName}
\title{Parameter names.}
\usage{
GetParName(dist, lang = "fr")
}
\arguments{
\item{dist}{character, distribution name}

\item{lang}{character, language ('en' or 'fr')}
}
\value{
A character vector.
}
\description{
Returns the names of the parameters of a distribution, 
in French (default) or English.
}
\examples{
GetParName('Normal')
GetParName('GEV')
GetParName('GEV',lang='en')
}
