% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_lib.R
\name{GetParNumber}
\alias{GetParNumber}
\title{Number of parameters.}
\usage{
GetParNumber(dist)
}
\arguments{
\item{dist}{character, distribution name}
}
\value{
An integer.
}
\description{
Returns the number of parameters of a distribution.
}
\examples{
GetParNumber('Normal')
GetParNumber('GEV')
}
