% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_lib.R
\name{GetPdf}
\alias{GetPdf}
\title{Probability Density Function (pdf)}
\usage{
GetPdf(y, dist, par, log = FALSE)
}
\arguments{
\item{y}{numeric, value at which the pdf is evaluated}

\item{dist}{character, distribution name}

\item{par}{numeric vector, parameter vector}

\item{log}{logical, returns log-pdf if TRUE}
}
\value{
The pdf or the log-pdf as a numeric.
}
\description{
Evaluates the pdf of a distribution
}
\examples{
GetPdf(0,'Normal',c(0,1))
GetPdf(200,'GEV',c(100,25,-0.2))
GetPdf(200,'GEV',c(100,25,0.2))
GetPdf(3,'Poisson',0.75)
}
