% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_lib.R
\name{GetQuantile}
\alias{GetQuantile}
\title{Quantile Function}
\usage{
GetQuantile(p, dist, par)
}
\arguments{
\item{p}{numeric in (0;1), nonexceedance probability}

\item{dist}{character, distribution name}

\item{par}{numeric vector, parameter vector}
}
\value{
The p-quantile as a numeric.
}
\description{
Evaluates the quantiles of a distribution
}
\examples{
GetQuantile(0.99,'Normal',c(0,1))
GetQuantile(0.99,'GEV',c(100,25,-0.2))
GetQuantile(0.99,'GEV',c(100,25,0.2))
GetQuantile(0.99,'Poisson',0.75)
}
