% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions_lib.R
\name{GetReducedVariate}
\alias{GetReducedVariate}
\title{Reduced variate}
\usage{
GetReducedVariate(p, dist)
}
\arguments{
\item{p}{numeric in (0;1), nonexceedance probability}

\item{dist}{character, distribution name}
}
\value{
The reduced variate with nonexceedance probability p.
}
\description{
Returns the 'reduced variate' that is used in some quantile plots 
(see e.g. quantile curve on Gumbel paper)
}
\examples{
GetReducedVariate(0.99,'Normal')
GetReducedVariate(0.99,'Gumbel')
GetReducedVariate(0.99,'GEV')
GetReducedVariate(0.99,'Poisson')
}
