% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{KS}
\alias{KS}
\title{Kolmogorov-Smirnov Test}
\usage{
KS(y, dist, par)
}
\arguments{
\item{y}{numeric vector, data}

\item{dist}{character, distribution name}

\item{par}{numeric vector, parameter vector}
}
\value{
A list with the following components:
    \item{pval}{numeric, p-value of the test}
    \item{stat}{numeric, test statistics}
    \item{xtra}{numeric, xtra information: empty for this test}
}
\description{
Applies a one-sample Kolmogorov-Smirnov test (see ?stats::ks.test)
}
\examples{
y=stats::rnorm(20)
KS(y,'Normal',c(0,1))
KS(y,'Normal',c(1,1))
KS(y,'Gumbel',c(0,1))
}
