% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmcoptions_def}
\alias{mcmcoptions_def}
\title{Default MCMC options}
\format{
An object of class \code{list} of length 10.
}
\usage{
mcmcoptions_def
}
\description{
A named list containing the default MCMC options.
See ?Hydro3_Estimation for more details.
}
\keyword{datasets}
