% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{MK}
\alias{MK}
\title{Mann-Kemdall Test}
\usage{
MK(y)
}
\arguments{
\item{y}{numeric vector, data}
}
\value{
A list with the following components:
    \item{pval}{numeric, p-value of the test}
    \item{stat}{numeric, test statistics}
    \item{xtra}{numeric, xtra information: empty for this test}
}
\description{
Applies the Mann-Kendall trend test
}
\examples{
y=stats::rnorm(50)
MK(y)
y=y+0.1*(1:length(y))
MK(y)
}
