% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{options_def}
\alias{options_def}
\title{Default estimation options}
\format{
An object of class \code{list} of length 9.
}
\usage{
options_def
}
\description{
A named list containing the default estimation options.
See ?Hydro3_Estimation for more details.
}
\keyword{datasets}
