% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{Pettitt}
\alias{Pettitt}
\title{Pettitt Test}
\usage{
Pettitt(y)
}
\arguments{
\item{y}{numeric vector, data}
}
\value{
A list with the following components:
    \item{pval}{numeric, p-value of the test}
    \item{stat}{numeric, test statistics}
    \item{xtra}{numeric, xtra information: position of the step change}
}
\description{
Applies the Pettitt step-change test
}
\examples{
y=stats::rnorm(50)
Pettitt(y)
y[26:50]=y[26:50]+2
Pettitt(y)
}
